Split the connection specs into separate files
